renamed slider toggle
